Fix bad detection of 64bit arch on x32
This commit comes from a Debian bug report from Thorsten:

  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=850723

The important part is:

+++++++++++++++++++++++++++++++++
As for libtommath… its build logs do not show this warning, but
that may be due to a difference in CFLAGS, and it does not have
such an extensive testsuite as heimdal, so it might also need to
have this patch applied (and forwarded upstream).

A better fix *might* be to fix MP_MASK to read something along
the lines of shifting 1ULL left, but the cast should already do
that, so I’m unsure why this doesn’t help. Do note that x32 is
an ILP32 architecture with 64-bit wide CPU registers (but 32-bit
pointers and longs), so 64-bit mode “should” work (and fast, at
that) but can be a bit tricky, and that the patch I attached is
positively known to fix several issues the heimdal testsuite shows.
+++++++++++++++++++++++++++++++++

This commit is a port of the patch proposed by Thorsten.