Prevent potential runtime-error in tests
`abs()` can only convert `INT_MIN-1 .. -1` to a positive `int`.
Nothing prevents the PRNG to create `INT_MIN` which then leads to a failure
of the call to `abs()` as seen in [0].
Instead add an unsigned version of the function reading from the PRNG,
so we also don't need to make an absolute value from it.

[0] https://github.com/libtom/libtommath/pull/555#issuecomment-1572004663

```
demo/test.c:1112:13: runtime error: negation of -2147483648 cannot be represented in type 'int'; cast to an unsigned type to negate this value to itself
```

Signed-off-by: Steffen Jaeckel <s@jaeckel.eu>